requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.9.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
h11==0.14.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.37
typing_extensions==4.9.0
tzdata==2024.1
uvloop==0.19.0
Werkzeug==3.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
